+++ ydb/core/formats/arrow/accessor/plain/accessor.cpp
@@ -14,7 +14,7 @@ std::optional<ui64> TTrivialArray::DoGetRawSize() const {
 std::vector<TChunkedArraySerialized> TTrivialArray::DoSplitBySizes(
     const TColumnLoader& saver, const TString& fullSerializedData, const std::vector<ui64>& splitSizes) {
     auto schema = std::make_shared<arrow::Schema>(arrow::FieldVector({ std::make_shared<arrow::Field>("f", GetDataType()) }));
-    auto chunks = NArrow::NSplitter::TSimpleSplitter(saver.GetSerializer())
+    auto chunks = NSplitter::TSimpleSplitter(saver.GetSerializer())
                       .SplitBySizes(arrow::RecordBatch::Make(schema, GetRecordsCount(), { Array }), fullSerializedData, splitSizes);
     std::vector<TChunkedArraySerialized> result;
     for (auto&& i : chunks) {
