+++ ydb/core/formats/arrow/accessor/sub_columns/columns_storage.cpp
@@ -1,0 +1,3 @@
+#include "columns_storage.h"
+
+namespace NKikimr::NArrow::NAccessor::NSubColumns {}   // namespace NKikimr::NArrow::NAccessor::NSubColumns
