+++ ydb/core/formats/arrow/accessor/sub_columns/iterators.cpp
@@ -1,0 +1,5 @@
+#include "iterators.h"
+
+namespace NKikimr::NArrow::NAccessor::NSubColumns {
+
+}   // namespace NKikimr::NArrow::NAccessor::NSubColumns
