+++ ydb/core/formats/arrow/accessor/sub_columns/settings.cpp
@@ -1,0 +1,5 @@
+#include "settings.h"
+
+namespace NKikimr::NArrow::NAccessor::NSubColumns {
+
+}   // namespace NKikimr::NArrow::NAccessor::NSubColumns
