+++ ydb/library/formats/arrow/protos/accessor.proto
@@ -34,12 +34,11 @@ message TConstructor {
 }
 
 message TSubColumnsAccessor {
-    message TSchema {
-        optional string Description = 1;
-    }
-    optional TSchema Schema = 1;
+    optional uint32 ColumnStatsSize = 1;
+    optional uint32 OtherStatsSize = 2;
     message TColumn {
-        optional string Description = 1;
+        optional uint32 Size = 1;
     }
-    repeated TColumn Columns = 2;
+    repeated TColumn Columns = 3;
+    optional TColumn Others = 4;
 }
